refactor: move and rename prisma.js into /utils/prismaClient.js
