+++ controllers/articleCommentController.js
@@ -1,4 +1,4 @@
-import prisma from '../prisma.js';
+import prisma from '../utils/prismaClient.js';
 import { assert } from 'superstruct';
 import { CreateArticleComment, PatchArtcleComment } from "../structs.js";
