+++ controllers/productCommentController.js
@@ -1,4 +1,4 @@
-import prisma from '../prisma.js';
+import prisma from '../utils/prismaClient.js';
 import { assert } from 'superstruct';
 import { CreateProductComment, PatchProductComment } from "../structs.js";
